feat: creating a new timer using requestAnimationFrame to persist the right time even when the user change the window on the browser
